+++ packages/v1/sdk-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/sdk-js
 
+## 1.53.1-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.53.1-next.0
+
 ## 1.53.0
 
 ### Patch Changes