+++ packages/v2/agent/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @copilotkitnext/agent
 
+## 1.53.1-next.0
+
 ## 1.53.0
 
 ### Patch Changes