+++ packages/v2/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkitnext/core
 
+## 1.53.1-next.0
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.53.1-next.0
+
 ## 1.53.0
 
 ### Patch Changes